Add files via upload
In short: 
a) Fernet generates a key, that key can be used to encrypt and decrypt data. It may look like this
   "9AFosuDE9Ytk4dPXUDhyPdMhus0_d5knBp585mLlGs4=""
b) Next, a blank dictionary should be made and saved to a file called "Vault.txt". Here is where passwords will be stored. in a format like 
	{"Netflix":"P@55w0rd!", "Amazon":"Password123!"}
c) The key will be used to encrypt and decrypt the data in Vault.
d) When not in use, the key should be zipped with a password (the master password). This will only be unzipped to decrypt or encrypt "Vault.txt"

So far, I made a function to generate passwords to populate Vault.txt: "generatePassword"
I added a function to zip and delete the Fernet Key. (It also adds password protection): "zipWithUserPass" 
I added functions required to test encoding through Fernet: "testFernet()"
these will be the core functions needed to operate the password management system. Hope this helped

TODO: make a function to unzip the key.key.zip file to grab the Fernet key to later decrypt the contents of Vault.txt. This should ONLY be
	   held in memory.